teamproject/backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
Django==5.1.3
django-cors-headers==4.6.0
django-livereload-server==0.5.1
djangorestframework==3.15.2
Flask==3.0.3
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.0
MarkupSafe==2.1.5
multidict==6.1.0
numpy==2.0.0
openai==0.28.0
packaging==24.2
pillow==11.0.0
propcache==0.2.0
pydantic==2.10.2
pydantic_core==2.27.1
pygame==2.6.0
pygame-ce==2.5.0
pygame_gui==0.6.12
pytesseract==0.3.13
python-dotenv==1.0.1
python-i18n==0.3.9
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.1
tornado==6.4.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.3
yarl==1.18.0
